backend/requirements.txt
Django==4.2.17
djangorestframework==3.14.0
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
Pillow==10.2.0  # For image handling
psycopg2-binary==2.9.9  # For PostgreSQL support
gunicorn==21.2.0  # For production deployment 